refactor(product): tier 2 usecase 구현
