--- src/order/application/order-stat.service.ts
@@ -1,14 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { GetPopularProductsUseCase } from "@/order/application/use-cases/tier-1-in-domain/get-popular-products.use-case";
-import { PopularProductResult } from "@/order/domain/interfaces/order-item.repository.interface";
-
-@Injectable()
-export class OrderStatApplicationService {
-  constructor(
-    private readonly getPopularProductsUseCase: GetPopularProductsUseCase
-  ) {}
-
-  async getPopularProducts(limit: number): Promise<PopularProductResult[]> {
-    return await this.getPopularProductsUseCase.execute({ limit });
-  }
-}
