+++ src/product/product.module.ts
@@ -6,7 +6,7 @@ import { ProductApplicationService } from "./application/services/product.servic
 import { ProductRepository } from "./infrastructure/persistence/product.repository";
 import { ProductTypeOrmEntity } from "./infrastructure/persistence/orm/product.typeorm.entity";
 import { GetProductByIdUseCase } from "./application/use-cases/tier-1-in-domain/get-product-by-id.use-case";
-import { GetProductByIdsUseCase } from "./application/use-cases/tier-1-in-domain/get-product-by-ids.use-case";
+import { GetProductsByIdsUseCase } from "./application/use-cases/tier-1-in-domain/get-products-by-ids.use-case";
 import { GetAllProductsUseCase } from "./application/use-cases/tier-1-in-domain/get-all-products.use-case";
 import { ReserveStockUseCase } from "./application/use-cases/tier-1-in-domain/reserve-stock.use-case";
 import { ReleaseStockUseCase } from "./application/use-cases/tier-1-in-domain/release-stock.use-case";
@@ -20,6 +20,7 @@ import { OrderModule } from "../order/order.module";
 import { ReserveStockDomainService } from "./domain/services/reserve-stock.service";
 import { ReleaseStockDomainService } from "./domain/services/release-stock.service";
 import { ConfirmStockDomainService } from "./domain/services/confirm-stock.service";
+import { GetPopularProductsWithDetailUseCase } from "./application/use-cases/tier-2/get-popular-products.use-case";
 
 @Module({
   imports: [
@@ -35,12 +36,13 @@ import { ConfirmStockDomainService } from "./domain/services/confirm-stock.servi
     TransactionService,
     ProductApplicationService,
     GetProductByIdUseCase,
-    GetProductByIdsUseCase,
+    GetProductsByIdsUseCase,
     GetAllProductsUseCase,
     ReserveStockUseCase,
     ReleaseStockUseCase,
     ConfirmStockUseCase,
     GetStockReservationsByKeyUseCase,
+    GetPopularProductsWithDetailUseCase,
     ReserveStockDomainService,
     ReleaseStockDomainService,
     ConfirmStockDomainService,
